remove unnecesary stuff
