+++ .env
@@ -1,0 +1,3 @@
+HOST=0.0.0.0
+PORT=8080
+ENVIRONMENT=development
