+++ go.mod
@@ -1,4 +1,4 @@
-module classconnect
+module user_service
 
 go 1.23.2
