--- service/state.go
@@ -1,6 +1,0 @@
-package main
-
-type State struct {
-	Courses   []Course
-	IdCounter uint64
-}
